requirements.txt
# =====================
# Core Web
# =====================
fastapi==0.115.12
uvicorn[standard]==0.35.0
python-dotenv==1.1.0
python-multipart==0.0.20

# =====================
# OpenAI / LangChain
# =====================
openai>=1.30.0,<2.0.0
langchain==1.2.9
langchain-core==1.2.9
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-huggingface==1.2.0
langchain-openai==1.1.7

# =====================
# Vector / PDF
# =====================
faiss-cpu==1.8.0
pypdf==4.2.0

# =====================
# ML / Audio
# =====================
numpy==1.26.4
protobuf==4.25.3
ctranslate2==4.6.0
faster-whisper==1.1.1
torch==2.6.0
onnxruntime==1.21.0
pydub==0.25.1
gTTS==2.5.4

# =====================
# HTTP / Async
# =====================
httpx==0.28.1
aiohttp==3.11.16

# =====================
# Auth / Utils
# =====================
python-jose[cryptography]==3.3.0
tiktoken==0.9.0


huggingface-hub>=0.24.0
transformers==4.41.2
packaging==23.2
wheel<0.46



